[implicit] introduce flux var cache
The flux variables do not exist as a global vector in the model
anymore. Instead, a cache class can be saved per scv face in
order to store data that should be recycled. This can be switched
on via the property EnableFluxVariablesCache and leads to a better
performance on small grids. On large grids memory usage becomes
an issue and it should be switched off.
If the property ConstantBoundaryConditions is switched on, the
volume variables at the boundary, used by cc models, will be stored
in the cache and recomputation does not occur.