[implicit][fluxVars] rename update function
The flux variables base class now has the init() routine
which takes care of setting the pointers and eventually
the stencil. The actual implementation of the flux vars
calls this method in its method initAndComputeFluxes().
This eventually precomputes the advective fluxes. Wherever
only the stencil is desired (localjacobian, stencilvector),
only the init() method from the base class is called in order
to save computational time.