[implicit] make vol vars cache optional
The volume variables caching can now be disabled in order for the
simulation to use less memory. The FVGeometry can be bound to an
element which then creates the geometries necessary for operations
on the element or the complete stencil. Several classes now need to
be friend with the model, otherwise the FVGeometry can not be bound
to the element. Maybe the bind function could return a geometry instead
of a reference, then access rights can be limited better. It would decrease
the efficiency though.