[implicit] introduce flags for discretization method
Each discretization method (Box, CCTpfa, Mpfa etc) now has a corresponding
flag which is set as a property. This allows a specialization of the used
constitutive laws for the different methods. For a model it is thus no longer
necessary to define additional typetags for the different discretization methods.
On the problem level one can simply derive from the discretization method's typetag
and the model type tag. Additional method-dependend classes are automatically set via
the DiscretizationMethod property in the form of template specializations