[geomechanics][elastic] implement elastic model in new structure
This commit introduces the necessary changes and new classes for a
cctpfa elastic model. Checking the results suggest, that a simple
tpfa scheme is not suitable for linear elasticity, not even on structured
grids. It is thus questionable wether or not this implementation will be deleted
in the future. Note that it only works in 2d so far.