[effectivepermeabilityrutqvisttsang] Outsourced permeability calculation
Reviewed by heck

Instead of calculating it in the local residual, the effective permeability
is now calculated within a new file permeabilityrutqvisttsang.hh in the
folder dumux/material/fluidmatrixinteractions.
This is just one possible effective permeability model. Several others can be
added and set via the property EffectivePermeabilityModel.

Up to now, the permeabilities were averaged twice, first by calculating
the mean of the mean of the intrinsic permeabilities in the flux variables
and later by averaging the effective permeabilities in the local residual.
This is redundant. Now, only the mean of the effective permeability is
calculated in the local residual.
This change required to change the reference files for the sequential and
the parallel test.