[implicit][assembler] remove assembleElement_() method
The assemble_() method in the ImplicitAssembler now calls the
assemble routine in the localjacobian directly. The global matrix
and the residual are passed to that routine where the residua and
derivatives are directly set in the global containers. That makes
a specialized assembleElement_() method obsolete as the local matrix
is no longer stored.