[implicit change caching property name and move files
This commit introduces two things:
1. Renaming of the enableCache properties to enableGlobalCache.
The cache classes will also be used to store temporary data in case the
caching is disabled. Thus the properties should indicate that they
switch on a GLOBAL caching. The commit further changes the corresponding
properties set in the defaults.
2. The volume variables vector is now specialized for the box and cc
methods. This commit thus also introduces the moving of the volvarvector
file into the cc folder and renaming of the class