introduce discretization folder
A new folder named discretization is introduced on the /dumux level.
In this folder there are all the geometry classes as well as vol vars, flux var cache
and physical laws (darcy, fick etc), since they could be possibly reused in decoupled
models and are not restricted to implicit models. NOTE: The fluxvariables and the flux
variables cache are still in the folder /porousmediumflow/implicit/. It is to be discussed
whether or not they should be moved into the discretization folder as well.