[fvGeometry] introduce global-local-bind concept
The global object can create an empty temporary local object that
can be bound to an element/stencil. This is closer to the
current dumux-stable implementation and still allows to switch
between global caching (for smaller problems where memory is not an issue)
and local caching (variables are precomputed for each element assemble temporarily)

Works for cc and box (in 2d). Box needs some performance improvement,
probably when creating the box geometries. But also other parts
might be possbile to optimize.