local objects always get const ref to global object
The local objects should not be getting a non const reference
to the corresponding global objects. This way we can get rid of the
list of friend classes in the model needing access to the non const
containers. The local jacobian needs to be friend of the global cache
containers and the model, as it does need to change vol vars during the
calculation of the derivatives. Depending on the global cache being
enabled or not, the corresponding reference to the vol var object
is now obtained in a new specialized method.

(cherry picked from commit 8482b8c9efac0022dd38acbc05424870d4ad27f4)