make subcontrolvolume entities base classes simple interfaces
The scv and scvf base classes now get the actual implementation
as a template parameter and serve as an interface for scv and scvf
implementations. Furthermore, the box scv and scvf does not store a geometry
anymore, but only the corner points. The geometry is built on the fly if requested.
The BoxGeometryHelper is adjusted to this new concept, as well as the global and
element-fvgeometry.

(cherry picked from commit 8b55b5d6db4414e57072ed05b8903a5ec49b0ee5)