[mpfa][fluxvarcache] fix update method
When additional mpfa methods are implemented, the types used for
vectors and stencil can differ for boundary and inner interaction volumes.
In the L-method for example, the sizes of vectors etc in the interaction
volumes are known, thus, statically sized types can be used. The updateInner
method has therefore been adapted to handle different types. The locally
saved types are always the dynamic ones to guarantee compatibility on the boundary.