[gridcreator] Fix gmsh parameter element index handling
We now get the parameter using the insertion index because that's the index
used by the factory in the gmsh reader. This also fixes how UG and ALU handle
reading gmsh files differently in parallel again.

Unfortunately the gmsh reader doesn't provide any load balancing features for the
parameters so that they might not always work in parallel

We now get the parameter using the insertion index because that's the index
used by the factory in the gmsh reader. This also fixes how UG and ALU handle
reading gmsh files differently in parallel again.

Unfortunately the gmsh reader doesn't provide any load balancing features for the
parameters so that they might not always work in parallel?