[common] Add a dof mapper that reorders the dofs optimizing the matrix sparsity pattern
The mapper computes a permutation of the index set using the external PT-SCOTCH library.
This is a simple replacement for element mapper (cc) / vertex mapper (box). It requires to
store on int per dof for the mapping (fast direct vector access).
This makes the matrix bandwidth small and reduces the fillins during LU factorization.
So it helps especially for direct solvers or iterative solver preconditions by ILU.
The reordering is particularly helpful for network problems with foamgrid to optimize
the matrix sparsity pattern.