[2pnc] Use mole fraction gradients instead of concentration gradients
This was wrongly altered in a previous commit. Fick's law depends on
the mole fraction gradient. The density has to be approximated at
the face.

Also fixes the diffusive fluxes that were sometimes wrongly added
to the total mass balance when using the replaceCompIdx to replace
one component balance by the total mass balance.