[fluxvariables] fix scaling of neumann fluxes in case of flux var caching
the neumann fluxes are divided by rho/mu to recover Kgradp during the set up
of the local systems in the interaction volumes. This factor is multiplied
again to the resulting flux in order to restore a mass flux for the corresponding
boundary face. In case of global caching the initial solution was used during the scaling.
This has to be used again here...