[implicit][propertydefaults] set AMG as default linear solver
The default linear solver is set to AMG. After changing this, some tests
failed. If the differences in the solution were in the order of machine
epsilon, the zero threshold value has been adjusted. In the remaining cases
the origin of the tests to fail after changing the linear solver could not be
detected and the original linear solver has been set as a property for these tests.