[gridcreator] Improve grid data handling
Gmsh data for Alugrid and ug needs to be manually loadbalances. Sadly, ug
doesn't implement parameter load balancing for elements at the moment (only vertices).
It might be possbile to have all parameters on all processes.

Use always level 0 entities. This is for user convenience. The data was always
only available for level 0. Now the gridcreator also checks this and if not we take
the level 0 father instead to evaluate the parameter - more robust.