[tpfa][globalfvgeometry] save copy to gridview instead of reference
This commit undos the part of the last commit introducing the references
to the gridview instead of copies. It was believed that the copy was the reason
why it failed for adaptive grids. Instead, it was only because of the local
indexing (fixed by the rest of the previous commit).