[mpfa] allow for different interaction volume seed types
Up to now the types for the inner and boundary interaction volume seeds
were always identical. For the mpfa-l and other methods, it would be more efficient
if these types could differ, as e.g. the sizes of the containers are known at compile time
for the l method. This commit introduces the capability of the types to differ.