[mpfa] make scv face method-specific
We introduce method-specific implementations that should inherit from a
mpfa-specific base class (which inherits from the scvf base class). We define
an interface for the constructor that has to be met by all the implementations.
We do this so that the classes instantiating scv faces do not have to be
method-specific. We add a local index to the constructor which corresponds
to the local vertex index an scvf is connected to for the mpfa-o-fps scheme.