[start] Refactor parameter parser to not need TypeTag
The parameter parser previously needed a TypeTag traits template argument only
to get the parameter tree type although a parameter tree is already handed it.
This was done to avoid an usused parameter message otherwise issued if the parameter
macros (which need the TypeTag) are not used. Twisted :/ (can't we refactor the parameter system?)

The new implementation has an optional input file name parameter which is taken from the
parametertree if it was provided as command line argument in the default start method. You
can also supply a specific file now statically in the code. If no name is given, the
parameter parser defaults to the <exectuable>.input pattern.