Merge branch 'feature/defaulttoamg' into 'master'
[implicit][propertydefaults] set AMG as default linear solver

The default linear solver is set to AMG. The solver is parallel in contrast to the current default sequential solver and
shows the best performance in a wide range of problems.

After changing this, some tests
failed. If the differences in the solution were in the order of machine
epsilon, the zero threshold value has been adjusted. 

In the remaining cases
the origin of the tests to fail after changing the linear solver could not be
detected and the original linear solver has been set as a property for these tests.
(it quite possibly has to be a bug in the models as the same jacobian should lead to the same solution,
the newton convergence criteria were tried in the most precise configuration and there were still differences)

TODO
* [x] test corner point grid

This fixes #325.

See merge request !218