[mpfa] fix some bugs for 2d grids in 3d
The function to make the indices unique in the local sub-control entity seeds
was buggy and led to wrong outside scv faces in the assembly of the local matrices.

Furthermore, calling scvfTouchesBranchingPoint() for grids with dim = dimWorld would have led
to a segmentation fault.

Also, in the interaction volume, when getting the transmissibilities, we always have to compute them
again when coming from the outside face, as the normal vector of the outside face can always be different to
that of the inside face, even when not on a branching point.