[mpfa][globalfvgeom] fix bug with numboundaryscvf
The boundary faces counter is now per scvf and not per intersection anymore
(which was bad anyway regarding the naming), thus, we do not need to multiply
it by two anymore at the end of creating the dual grid geometries. We originally
counted it only once since two scvfs on the same intersection had the same outside vol
var index. Now there are boundary volume variables per scvf (better approximation of the
boundary).