[1p2c] port ni models to new structure
Note that the niconvectionproblem has been ported before. That implementation
had in mind that an interface change of the spatial params might occur in the future
which does not allow a simple update of the volume variables. This commit brings it
back to the old approach. If the spatial parameter interface change happens we'll have to
change all the models anyway.