[1p] port ni tests to new interface
The tests fail, but they do so because the output has changed. This has
to be checked again.[1p] port ni tests to new interface

The tests fail, but they do so because the output has changed. This has
to be checked again.[1p] port ni tests to new interface

The tests fail, but they do so because the output has changed. This has
to be checked again.[1p] port ni tests to new interface

The tests fail, but they do so because the output has changed. This has
to be checked again.[1p] port ni tests to new interface

The tests fail, but they do so because the output has changed. This has
to be checked again.[1p] port ni tests to new interface

The tests fail, but they do so because the output has changed. This has
to be checked again.[1p] port ni tests to new interface

The tests fail, but they do so because the output has changed. This has
to be checked again.[1p] port ni tests to new interface

The tests fail, but they do so because the output has changed. This has
to be checked again.