[mpfa-o][localscvf] store pointer to scvf instead of copies
The interaction volumes have to be instantiated with a valid fvGeometry etc and
during assembly they only exist temporarily for the transmissibility calculations.
Interaction volumes can now not be instantiated and stored globally anymore, but, why
would one like to do that. If global flux var caching is enabled, then the transmissibilities
are stored over the whole simulation time. This is the main thing one wants from them.