[stencil] Improve flux stencils
Fix the flux stencil for every method independent of the stencils used
in the different flux laws. The matrix entries have to be reserved for all
possible occuring stencils if the method has parameter or solution dependent
stencils, otherwise the matrix needs to be rebuild every time. This implements
this unified flux stencil for box, cc, ccmpfa.
Remove a lot of duplicate code.
Separate the physical laws from the dicretization stencil.