[mpfa][iavol] make other mpfa methods using o-method on the boundaries work again
We now call the Implementation from the o-method interaction volume because we need
to incorporate some terms in the problems with facet coupling.
Todo: Problems using interior interaction volumes that inherit from the o-method
specialization AND use the o-method on the boundaries do not work at the moment. We
would have to tell the o-method the actual implementation somehow introducing a new
template parameter for the implementation.