[mpfa-o] interaction volume gets implementation as template parameter
by passing the implementation of the actual method's interaction volume type
to the o-method interaction volume we are able to overload functionalities in
derived classes and still use different interaction volume types on the boundary and
inside the domain. This was not possible through the property system as it was unclear
which interaction volume is the actual implementation.