[mixeddimension][facetcoupling] remove use of std::pair in coupling data
We no longer return a pair of bool and element index when obtaining the
coupled low dim element for a given bulk scvf. Instead the low dim index
is returned directly, otherwise an exception is thrown.
Also, we no longer use pairs to store the element indices and corresponding
scvf indices of the bulk elements that directly couple to a low dim element.
We store these data in two separate vectors to make them accessible independently.