[basicproperties] Introduce new basic properties NumEqVector and PrimaryVariables
We need to distinguish between two concept for sake of generality:
* a vector of size number equations that we can use to store fluxes, storage, source and residuals for each
equation and:
* a primary variables vector that contains the values of all primary variables
at a d.o.f. location.
The latter might also have a state in the case of privar switch models
that indicates the type of primary variables saved in this vector.
The vectors have different usages but will default to the same type Dune::FieldVector<Scalar, numEq>
and can be spezialised independent from each other if needed.