[fem][model] add dof mapper return function
this is needed in order to compile because AMG calls this function.
However, right now we return the vertex mapper here which only works
for order = 1. For higher order, we have to implement corresponding mappers.
Or, depending on what AMG does, is it always admissible to return the vertex mapper?