[mpfa] change iv index set concept
the dual grid index set now does not contain the data on the global scv indices
that are connected to the faces. The local data is only available in the iv index set.
This should be better for methods other than the mpfa-o method, as in those the local mappings
of the dual grid index set is maybe useless (e.g. for the mpfa-l scheme).