[2p][adaptive] use right privars for vol vars update
For cell-centered scheme, the primary variables in the reconstruction
were not divided by the number of children that composed the primary
variables in a father element. This does not have an effect here as the
helper class and the tests only consider incompressible fluids. Furthermore,
the problem would only appears if in one refinement some elements get refined more
than one level. However, for the sake of readability it should be correct here.