[2padaptive] Fix pressure reconstruction for 2p adaptionhelper
For some grids, e.g. UGGrid, the father element
is identical to the son element.
It is therefore not a good idea to use the
adaptedValue containers for the father and
for the son element since they might be also identical.
This has been changes such that the adaptedValues
are no longer modified in the reconstruction routine.