[mpfa] template classes by interaction volume instead of MpfaMethod
We now do not specialize the interaction volume depending on the method chosen
but extract the chosen method from the interaction volume type that has been
set as property. This enables us to hand in the traits class to the interaction
volume from the outside, which is actually what you want. Now, for certain grids
one can use static matrix/vector types to speed up simulations.