[pointsources] Do not enable by default
The point source computation in FVProblem uses the problem implementation to compute the point source map.
However it is executed in the constructor where the problem implemenation is not fully instatiated yet as it
derives from FVProblem. This patch disables point source computation by default. If you want to specify point
sources you have to manually call problem->computePointSourceMap() now.