[bboxtree] Improve epsilon computation for bbox/bbox intersection
In case one dimension is very small take the maximum between the to be tested box dimensions
in the corresponding axis direction. There is one corner case left where the epsilon is very small:
If there are two bboxes of lines, both parallel in an axis direction, their collision is only
detected if the coordinates are an exact match. However, we currently do not see any used case
where this might occur.