[mpfa][fluxcachefiller] outsource data handle preparation
This introduces another set of functions specialized via enable_if,
but this is necessary as otherwise compilation fails for the tracer model,
because it is tried to access pressures via the vol vars, which the volvars
of the tracer model do not provide. By ensuring that this is only called
for processes using mpfa, we are on the safe side.