[el2p] Use base class quantities in the NewtonController
The `ElTwoPNewtonController` defined several data members that are
already present in the base `NewtonController`. Delete these
members and use the corresponding base class members.

The communication call in the function `globalResidual` of the
`ElTwoPBaseModel` didn't compile. Comment this call such that the
function works at least for sequential runs. This allows to use
the residual criterion in the Newton solver. Throw if the routine
is called in a parallel run.