[gridcreator] add boundary marker load balancing for alu, more tests
Realize boundary marker load balancing for ALUGrid. To achieve this, adapt
the data handle such that also the boundary markers are balanced. Since
the data handle works on faces rather than intersections, transfer first
the vector that is given in terms of boundary segments to a vector holding
a parameter value for each mesh face. After balancing, transfer the data
back to a boundary segment vector.

Restructure the tests and add more tests. Both UG and ALU are now tested
for Gmsh 2/3d and DGF 2d. The ALU tests for Gmsh 3d still need to be
worked on.