[2p][1pnc][richardsnc] Improve naming / definition of phases and components and their indices
Fluid systems exported phase indices for wetting and nonwetting components. The fluid system
is not the right place to decide about wettability. This introduces wettingPhase() in the spatial
params which returns the index of the phase in the fluidsystem we consider the wetting phase.
This way wettability can change temporally and spatially.
Indices throughout the 2p/2p2c/2pnc nc and mpnc where renamed accordingly.