[box][geometryhelper] use ref element for boundary scvf corner determination
The corner determination for boundary scvfs was wrong when using ug or alu grid,
because only information directly from the intersection geometry was used. We now
obtain the corners by using the positions obtained via the reference element. In
order to do so, we simply store a reference to the reference element during construction
of the helper class.