[2pnc][test] Fix fuelcell test
What was broken:
The test used philtophob law which used the wrong residual saturations. It went into efftoabslaw with
the non-wetting saturation so swr and snr where the wrong way around. It also overwrote the krw/krn function
of the material law in use with a quadratic/cubic law. It also defined the capillary pressure negative but then
it would be pc = pw-pn.

What was fixed:
Use the wettingPhase function instead to make the gas phase the wetting phase

This changed:
- the pw solution as swr and snr were corrected (the wrong way around pw was equal)
- the pc solution because it is now positite
- the mobilities as they didn't use the real van genuchten curves