[io][vkt] p1 non-conforming vtk function for enriched vertex dofs
We make the p1 non-conforming vtk function dependent on a provided
mapper type. If the mapper type is a
Dune::MultipleCodimMultipleGeometryType mapper, we assume the data to be
organised such that for each element the data is stored in a vector with
element-local vertex ordering. If another mapper is provided, we assume
that data to be organised in a standard solution vector to which the
provided mapper maps adequately. If another mapper is provided, we
assume that data to be organised in a standard solution vector to which
the provided mapper maps adequately.