[io][vtkoutput] let velocityOutput decide number of phase velocities
This commit will enable the geomechanics framework to not have to
state the number of phases in geomechanical model traits. Also, when
velocity output for the geomechanical models is implemented, we will
have to distinguish in the vtk output module anyway, as then we would
have to call numSolidPhases(). Outsourcing it directly to the velocity
output seems like an elegant solution.