[geomechanics] remove cc-compatibility
Some of the interfaces were designed such that they are in agreement
with the porous medium flow framework where we make them work for
both cell-centered as well as the box scheme. In the geomechanics
framework we now make use of the fact that cell-centered schemes are not
considered and use more FEM-like interfaces for the parameters required
for stress evaluations. This might enable us to rename/redefine these
interfaces to work with both FEM as well as Box. However, if
cell-centered schemes for geomechanics are to be implemented one day,
the whole structure would have to be overthought.