[multidomain][facet] factorize coupling mapper and introduce base class
We now specialize the coupling mappers depending on the discretization
method used in the bulk domain. Update of the map occurs in the
implementations for the different discretization schemes, where the
update_ function in the base class can be used with an execution policy
that defines how entries in the coupling maps are added for a given lower-
dimensional element and the corresponding embedments. This commit
furthermore introduces the coupling mapper for the box scheme used in
the bulk domain.