[cornerpoint] update CpGridCreator to Opm 2018.04
It is not possible to ask for the employed Opm version in a
convenient way. Therefore, support for older versions has to be
dropped.

Adapt an include. Get rid of the `TypeTag` template parameter.
Perform `loadBlance` only in a parallel run. Otherwise, the
boundary segment indices are lost. Add a small workaround for the
`static Deck` to prevent segfaults.
Provide a patch for opm-grid, delete patches for older versions.
Update `patches/README`.