[test][gridcreator] convert parallel output to sequential vtk files
The distribution of a grid over several cores can be system-dependent,
rendering the direct comparison of resulting parallel vtk files
impossible. To enable a comparison, convert the parallel files into
sequential ones in the fuzzycompare Python script. With this approach,
it is possible to compare the parallel output against the sequential
one. Adapt the gridcreator tests correspondingly.

Concerning the converted vtk files, a few difficulties had to be
handled:
* The files usually have duplicated vertices at the process borders.
  Eliminate the duplicated entries.

* The files can have an empty `PointData` section yielding to problems
  since the number of children differs. Eliminate the empty section
  from the files.

* The attributes may be different for each node: the
  `NumberOfComponents` isn't stated explicity if it is equal to `1`.
  On the other hand, the minimum and maximum values are provided
  which is not the case for the original vtk output. Eliminate the
  check of the attributes if a conversion from pvd has happened.

Enhance the gridcreator tests using boundary markers with an
additional synchronization step for those markers in the parallel
regime.