[io][gridmanager] Implement grid manager and grid data classes without singleton
Implement a grid manager class that creates grids and a grid data object. The new class
is no longer a singleton. It is instantiated by the user. It is possible to get a
shared_ptr to the grid data to pass it on to e.g. problem / spatial parameter classes.

* No longer sets the overlap automatically (overlap to be checked by the grid geometry)
* Does not depend on discretization method or TypeTag anymore

Joint work with Kilian Weishaupt <kilian.weishaupt@iws.uni-stuttgart.de>