[io] add capability to restart from a VTK file
Add the TinyXML-2 parser. Add functions in the `Restart` class
that read from a given VTK file into a `SolutionVector`. The
names of the primary variables in the VTK file are handed over by
the problem. Adapt `applyInitialSolution` of the FVProblem such
that it invokes the restart functionality if a start time larger
than 0 is handed over.