[multidomain][facet] allow for more flexibility w.r.t domain markers
The coupling manager can now return the element a bulk scvf is
coinciding with. This element can be used to obtain the domain marker of
the lowdim element by accessing the lowdim problem via the coupling
manager and obtaining the marker from the spatial params. This can be
used to set different coupling conditions (Dirichlet/Neumann) depending
on the marker. Additionally, we allow extraction of a sub-domain
specific grid data object from the wrapper. This object can then be
passed directly to the sub-domain's spatial parameters.