[parallel] perform parallel tests also for UGGrid and ALUGrid
Also fix the test itself in the following sense. By default, in the
box method, material law parameters are evaluated at scv centers,
while primary variables are evaluated at vertices. If no interface
condition is used, the secondary variables handed to the Vtk
writer might now differ depending on the order in which the scvs
are evaluated for each vertex. In order to saty consistent, this
test evaluates the material law parameters at the vertex position.