[io][loadsolution] Make privarnames functions consistent
* We require the signature std::string(int, int) signature for all pv
names functions passed to load solution. You can still wrap a function
that just takes a privar index into a lambda with these two arguments.

* We make FluidSystem an optional template parameter for all functions
that don't make use of it. This way they can be also used directly in
loadSolution when not using the convenience free function pvNames.

* We return a function from the convenience free function now to be able
to include the param group