fix determination of fracture facets, introduce test on triangles
We reuse the grid creation facility of the facet coupling framework and
require grid files that have the information on the fracture facets in
them. This enables us to reuse the codimension one grid adapter of the
facet coupling framework to identify which matrix element facets
coincide with a fracture.

This commit also introduces a test on the same geometry as in the 2.12
release.

TODO: add test for quads and tetrahedra