[discretization] specialize evalSolution for primary variables with state
Interpolating the entries of primary variable vectors by means of shape
functions can be wrong: If the vectors have different states, the entries
can be associated with different physical fields. Since the function
`evalSolution` is physics-agnostic, this presents a simple solution
without adding knowledge about the physics. If the states are different
over all corners of the associated geometry, `evalSolution` returns
the primary variables associated with the vertex that is closest to
the evaluation point.