[io][gmsh][uggrid] fix reading a Gmsh file into a parallel UGGrid
Depending on the Dune version, the boundary markers are present on
all processes (<= 2.6) or on the root process only (>= 2.7). Try to
handle this in a flexible way in the corresponding data handle:
Determine if the minimum size over all processes of the boundary
markers vector is zero. If yes, assume that the root process contains
all markers and broadcast them.

Currently, it isn't possible to refine a parallel UGGrid that has
been read by a GmshReader, see dune-grid issue #83. Therefore, don't
refine in the corresponding test.