[test][freeflow] Call postTimeStep before timeLoop->advanceTimeStep
In postTimeStep the L2 error is calculated using a difference to the analytical solution.
Without this commit, the L2 error is calculated using a difference between the numerical solution
at a certain time and the analytical solution one time step ahead.

With this commit, the numerical and analytical solution
used to calculate the L2 error should correspond to the same time.